+++ src/physical_plan.rs
@@ -15,20 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use datafusion::execution::context::TaskContext;
-use datafusion::execution::runtime_env::RuntimeEnv;
-use datafusion::physical_plan::{displayable, ExecutionPlan, SendableRecordBatchStream};
-use datafusion_common::Result;
-use futures::StreamExt;
-use std::collections::HashMap;
+use datafusion::physical_plan::{displayable, ExecutionPlan};
 use std::sync::Arc;
 
-use crate::errors::py_datafusion_err;
-use crate::record_batch::PyRecordBatch;
-use crate::utils::wait_for_future;
 use pyo3::prelude::*;
-use tokio::runtime::Runtime;
-use tokio::task::JoinHandle;
 
 #[pyclass(name = "ExecutionPlan", module = "datafusion", subclass)]
 #[derive(Debug, Clone)]
@@ -63,24 +53,6 @@ impl PyExecutionPlan {
         let d = displayable(self.plan.as_ref());
         format!("{}", d.indent())
     }
-
-    pub fn execute(&self, part: usize, py: Python) -> PyResult<PyRecordBatchStream> {
-        let ctx = Arc::new(TaskContext::new(
-            "task_id".to_string(),
-            "session_id".to_string(),
-            HashMap::new(),
-            HashMap::new(),
-            HashMap::new(),
-            Arc::new(RuntimeEnv::default()),
-        ));
-        // create a Tokio runtime to run the async code
-        let rt = Runtime::new().unwrap();
-        let plan = self.plan.clone();
-        let fut: JoinHandle<Result<SendableRecordBatchStream>> =
-            rt.spawn(async move { plan.execute(part, ctx) });
-        let stream = wait_for_future(py, fut).map_err(|e| py_datafusion_err(e))?;
-        Ok(PyRecordBatchStream::new(stream?))
-    }
 }
 
 impl From<PyExecutionPlan> for Arc<dyn ExecutionPlan> {
@@ -92,25 +64,5 @@ impl From<PyExecutionPlan> for Arc<dyn ExecutionPlan> {
 impl From<Arc<dyn ExecutionPlan>> for PyExecutionPlan {
     fn from(plan: Arc<dyn ExecutionPlan>) -> PyExecutionPlan {
         PyExecutionPlan { plan: plan.clone() }
-    }
-}
-
-#[pyclass(name = "RecordBatchStream", module = "datafusion", subclass)]
-pub struct PyRecordBatchStream {
-    stream: SendableRecordBatchStream,
-}
-
-impl PyRecordBatchStream {
-    fn new(stream: SendableRecordBatchStream) -> Self {
-        Self { stream }
-    }
-}
-
-#[pymethods]
-impl PyRecordBatchStream {
-    fn next(&mut self, py: Python) -> PyResult<PyRecordBatch> {
-        let result = self.stream.next();
-        let batch = wait_for_future(py, result).unwrap()?; //.map_err(DataFusionError::from)?;
-        Ok(batch.into())
     }
 }
